refactor(mcp-server): split annotation conformance tests
