+++ crates/assay-mcp-server/src/proxy/annotation_conformance.rs
@@ -196,553 +196,4 @@ pub fn build_tool_annotation_conformance_record(
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use serde_json::{json, Value};
-
-    fn github_args() -> Value {
-        json!({"owner": "acme", "repo": "prod-app", "title": "ci-key"})
-    }
-
-    fn workspace_args() -> Value {
-        json!({"workspace_id": "acme", "principal": "alice@example.com"})
-    }
-
-    #[test]
-    fn declared_read_only_true_mismatches_observed_mutating() {
-        let rec = conformance_for(
-            &DeclaredToolAnnotations {
-                read_only: Some(true),
-                destructive: None,
-                idempotent: None,
-                open_world: None,
-            },
-            "github.add_deploy_key",
-            &github_args(),
-        );
-
-        assert_eq!(rec["schema"], json!(TOOL_ANNOTATION_CONFORMANCE_SCHEMA));
-        assert_eq!(rec["conformance"], json!("mismatched"));
-        assert_eq!(
-            rec["mismatch_kind"],
-            json!("declared_read_only_observed_mutating")
-        );
-        assert_eq!(rec["assessed_axes"], json!(["read_only"]));
-    }
-
-    #[test]
-    fn declared_non_destructive_mismatches_observed_destructive() {
-        let rec = conformance_for(
-            &DeclaredToolAnnotations {
-                read_only: Some(false),
-                destructive: Some(false),
-                idempotent: None,
-                open_world: None,
-            },
-            "workspace.modify_org_policy",
-            &workspace_args(),
-        );
-
-        assert_eq!(rec["conformance"], json!("mismatched"));
-        assert_eq!(
-            rec["mismatch_kind"],
-            json!("declared_non_destructive_observed_destructive")
-        );
-        assert_eq!(rec["observed"]["behavior_class"], json!("destructive"));
-    }
-
-    #[test]
-    fn compatible_assessed_hints_are_consistent_not_certified() {
-        let rec = conformance_for(
-            &DeclaredToolAnnotations {
-                read_only: Some(false),
-                destructive: Some(false),
-                idempotent: None,
-                open_world: None,
-            },
-            "github.add_deploy_key",
-            &github_args(),
-        );
-
-        assert_eq!(rec["conformance"], json!("consistent"));
-        assert_eq!(rec["mismatch_kind"], Value::Null);
-        assert_eq!(rec["observed"]["behavior_class"], json!("mutating"));
-        assert!(rec["non_claims"]
-            .as_array()
-            .unwrap()
-            .iter()
-            .any(|v| { v.as_str().unwrap().contains("consistent does not certify") }));
-    }
-
-    #[test]
-    fn missing_assessed_hints_are_undeclared_not_consistent() {
-        let rec = conformance_for(
-            &DeclaredToolAnnotations {
-                read_only: None,
-                destructive: None,
-                idempotent: Some(true),
-                open_world: Some(false),
-            },
-            "github.add_deploy_key",
-            &github_args(),
-        );
-
-        assert_eq!(rec["conformance"], json!("undeclared"));
-        assert_eq!(rec["assessed_axes"], json!([]));
-        assert_eq!(rec["declared"]["idempotent"], json!(true));
-        assert_eq!(rec["declared"]["open_world"], json!(false));
-        assert_eq!(rec["unassessed_axes"], json!(["idempotent", "open_world"]));
-    }
-
-    #[test]
-    fn unclassified_or_incomplete_calls_are_inconclusive() {
-        for (tool, args) in [
-            ("unknown.tool", json!({})),
-            ("github.add_deploy_key", json!({"owner": "acme"})),
-        ] {
-            let rec = conformance_for(
-                &DeclaredToolAnnotations {
-                    read_only: Some(true),
-                    destructive: Some(false),
-                    idempotent: None,
-                    open_world: None,
-                },
-                tool,
-                &args,
-            );
-
-            assert_eq!(rec["conformance"], json!("inconclusive"));
-            assert_eq!(rec["mismatch_kind"], Value::Null);
-            assert_eq!(rec["assessed_axes"], json!([]));
-        }
-    }
-
-    #[test]
-    fn record_has_no_verdict_delivery_or_sensitive_identity_fields() {
-        let rec = conformance_for(
-            &DeclaredToolAnnotations {
-                read_only: Some(false),
-                destructive: Some(false),
-                idempotent: None,
-                open_world: None,
-            },
-            "github.add_deploy_key",
-            &github_args(),
-        );
-
-        let text = serde_json::to_string(&rec).unwrap();
-        for forbidden in [
-            "decision",
-            "reason",
-            "forwarded",
-            "delivered",
-            "credential_alias",
-            "scopes",
-            "target_digest",
-            "caller_id",
-        ] {
-            assert!(
-                rec.get(forbidden).is_none(),
-                "annotation conformance record must not carry field {forbidden}"
-            );
-        }
-        assert!(
-            !text.contains("ci-key"),
-            "raw sensitive argument values must not be copied into the record"
-        );
-    }
-
-    // ---- Declared-annotation extraction (Increment 5b-1) --------------------------------------
-
-    #[test]
-    fn extract_reads_declared_boolean_hints() {
-        let declared = extract_declared_annotations(&json!({
-            "readOnlyHint": true,
-            "destructiveHint": false,
-            "idempotentHint": true,
-            "openWorldHint": false,
-        }));
-        assert_eq!(declared.read_only, Some(true));
-        assert_eq!(declared.destructive, Some(false));
-        assert_eq!(declared.idempotent, Some(true));
-        assert_eq!(declared.open_world, Some(false));
-    }
-
-    #[test]
-    fn extract_absent_hints_are_none_not_default() {
-        // MCP defines schema defaults (e.g. readOnlyHint false), but the carrier records what the
-        // server actually declared: an absent hint is undeclared, never the default.
-        let declared = extract_declared_annotations(&json!({"title": "Add deploy key"}));
-        assert_eq!(declared.read_only, None);
-        assert_eq!(declared.destructive, None);
-        assert_eq!(declared.idempotent, None);
-        assert_eq!(declared.open_world, None);
-    }
-
-    #[test]
-    fn extract_null_or_non_object_annotations_are_all_none() {
-        for value in [
-            json!(null),
-            json!("readOnlyHint"),
-            json!(["readOnlyHint"]),
-            json!(42),
-        ] {
-            let declared = extract_declared_annotations(&value);
-            assert_eq!(
-                (
-                    declared.read_only,
-                    declared.destructive,
-                    declared.idempotent,
-                    declared.open_world
-                ),
-                (None, None, None, None),
-                "non-object annotations must not declare any hint: {value}"
-            );
-        }
-    }
-
-    #[test]
-    fn extract_non_boolean_hint_values_are_none() {
-        // A hint declared with a non-boolean value is not a trustworthy boolean; record None.
-        let declared = extract_declared_annotations(&json!({
-            "readOnlyHint": "true",
-            "destructiveHint": 1,
-            "idempotentHint": null,
-            "openWorldHint": {},
-        }));
-        assert_eq!(
-            (
-                declared.read_only,
-                declared.destructive,
-                declared.idempotent,
-                declared.open_world
-            ),
-            (None, None, None, None)
-        );
-    }
-
-    #[test]
-    fn extract_partial_hints_keep_declared_and_drop_absent() {
-        let declared = extract_declared_annotations(&json!({"readOnlyHint": false}));
-        assert_eq!(declared.read_only, Some(false));
-        assert_eq!(declared.destructive, None);
-        assert_eq!(declared.idempotent, None);
-        assert_eq!(declared.open_world, None);
-    }
-
-    // ---- Observation basis and verb-sync (Increment 5b) ---------------------------------------
-
-    #[test]
-    fn incomplete_basis_forces_inconclusive_with_null_declared_and_digest() {
-        // Declared hints and digest are nulled (annotations were not observed), but the classifier's
-        // observed behavior is still recorded; conformance is forced inconclusive, never undeclared.
-        let rec = build_tool_annotation_conformance_record(
-            ObservationBasis::Incomplete,
-            &DeclaredToolAnnotations {
-                read_only: Some(true),
-                destructive: Some(false),
-                idempotent: None,
-                open_world: None,
-            },
-            "github.add_deploy_key",
-            Some("sha256:ignored-when-incomplete"),
-            &json!({"owner": "acme", "repo": "prod-app"}),
-        );
-        assert_eq!(rec["observation_basis"], json!("incomplete"));
-        assert_eq!(rec["conformance"], json!("inconclusive"));
-        assert_eq!(rec["mismatch_kind"], Value::Null);
-        assert_eq!(rec["assessed_axes"], json!([]));
-        assert_eq!(rec["tool"]["tool_digest"], Value::Null);
-        assert_eq!(rec["declared"]["read_only"], Value::Null);
-        assert_eq!(rec["declared"]["destructive"], Value::Null);
-        assert_eq!(rec["observed"]["behavior_class"], json!("mutating"));
-    }
-
-    #[test]
-    fn complete_basis_records_observation_basis_and_digest() {
-        let rec = build_tool_annotation_conformance_record(
-            ObservationBasis::Complete,
-            &DeclaredToolAnnotations {
-                read_only: Some(true),
-                destructive: None,
-                idempotent: None,
-                open_world: None,
-            },
-            "github.add_deploy_key",
-            Some("sha256:abc"),
-            &json!({"owner": "acme", "repo": "prod-app"}),
-        );
-        assert_eq!(rec["observation_basis"], json!("complete"));
-        assert_eq!(rec["tool"]["tool_digest"], json!("sha256:abc"));
-        assert_eq!(rec["conformance"], json!("mismatched"));
-    }
-
-    #[test]
-    fn every_classifier_verb_maps_to_an_observed_behavior() {
-        // observed_behavior() must stay in sync with the verbs tool_decision::classify can emit.
-        // classify currently emits only privileged mutating/destructive verbs, so each must map; a
-        // new privileged verb added there without a mapping would silently downgrade a contradiction
-        // to inconclusive. A non-mutating verb, if ever added, maps to None by design.
-        let calls = [
-            (
-                "github.add_deploy_key",
-                json!({"owner": "acme", "repo": "prod-app"}),
-            ),
-            (
-                "slack.add_member",
-                json!({"workspace_id": "acme", "user_id": "u1"}),
-            ),
-            (
-                "workspace.grant_admin",
-                json!({"workspace_id": "acme", "principal": "p"}),
-            ),
-            (
-                "workspace.change_role",
-                json!({"workspace_id": "acme", "principal": "p"}),
-            ),
-            (
-                "workspace.invite_external",
-                json!({"workspace_id": "acme", "principal": "p"}),
-            ),
-            (
-                "workspace.modify_org_policy",
-                json!({"workspace_id": "acme", "principal": "p"}),
-            ),
-            (
-                "workspace.create_workspace_token",
-                json!({"workspace_id": "acme", "principal": "p"}),
-            ),
-        ];
-        for (tool, args) in calls {
-            let classified = classify(tool, &args);
-            assert_eq!(classified.state, "classified", "{tool} should classify");
-            assert!(
-                observed_behavior(classified.verb).is_some(),
-                "verb {:?} from {tool} has no observed_behavior mapping",
-                classified.verb
-            );
-        }
-    }
-
-    // ---- Shared producer/consumer contract fixture (Increment 5) ------------------------------
-    //
-    // The canonical `assay.tool_annotation_conformance.v0` contract is REAL output of
-    // `build_tool_annotation_conformance_record`, not a hand-authored mirror. Plimsoll will vendor
-    // the same file in a later slice and assert its consumer accepts each record.
-    // Regenerate after an intentional producer change: ASSAY_UPDATE_GOLDEN=1 cargo test -p
-    // assay-mcp-server tool_annotation_conformance_contract_fixture.
-
-    fn contract_fixture_path() -> std::path::PathBuf {
-        std::path::PathBuf::from(env!("CARGO_MANIFEST_DIR"))
-            .join("tests/fixtures/tool_annotation_conformance_contract.v0.json")
-    }
-
-    type ContractCase = (
-        &'static str,
-        ObservationBasis,
-        DeclaredToolAnnotations,
-        &'static str,
-        Value,
-        Option<&'static str>,
-    );
-
-    fn contract_records() -> Vec<Value> {
-        let cases: Vec<ContractCase> = vec![
-            (
-                "consistent_read_only_false_additive",
-                ObservationBasis::Complete,
-                DeclaredToolAnnotations {
-                    read_only: Some(false),
-                    destructive: None,
-                    idempotent: None,
-                    open_world: None,
-                },
-                "github.add_deploy_key",
-                json!({"owner": "acme", "repo": "prod-app", "title": "ci-key"}),
-                Some("sha256:tooldigest-consistent-readonly-false"),
-            ),
-            (
-                "consistent_destructive_false_additive",
-                ObservationBasis::Complete,
-                DeclaredToolAnnotations {
-                    read_only: Some(false),
-                    destructive: Some(false),
-                    idempotent: None,
-                    open_world: None,
-                },
-                "github.add_deploy_key",
-                json!({"owner": "acme", "repo": "prod-app"}),
-                Some("sha256:tooldigest-consistent-nondestructive"),
-            ),
-            (
-                "mismatched_read_only_mutating",
-                ObservationBasis::Complete,
-                DeclaredToolAnnotations {
-                    read_only: Some(true),
-                    destructive: None,
-                    idempotent: None,
-                    open_world: None,
-                },
-                "github.add_deploy_key",
-                json!({"owner": "acme", "repo": "prod-app"}),
-                Some("sha256:tooldigest-readonly-mismatch"),
-            ),
-            (
-                "mismatched_non_destructive_destructive",
-                ObservationBasis::Complete,
-                DeclaredToolAnnotations {
-                    read_only: Some(false),
-                    destructive: Some(false),
-                    idempotent: None,
-                    open_world: None,
-                },
-                "workspace.modify_org_policy",
-                json!({"workspace_id": "acme", "principal": "alice@example.com"}),
-                Some("sha256:tooldigest-destructive-mismatch"),
-            ),
-            (
-                "undeclared",
-                ObservationBasis::Complete,
-                DeclaredToolAnnotations {
-                    read_only: None,
-                    destructive: None,
-                    idempotent: None,
-                    open_world: None,
-                },
-                "github.add_deploy_key",
-                json!({"owner": "acme", "repo": "prod-app"}),
-                Some("sha256:tooldigest-undeclared"),
-            ),
-            (
-                "inconclusive_unknown_tool",
-                ObservationBasis::Complete,
-                DeclaredToolAnnotations {
-                    read_only: Some(true),
-                    destructive: Some(false),
-                    idempotent: None,
-                    open_world: None,
-                },
-                "unknown.tool",
-                json!({}),
-                Some("sha256:tooldigest-unknown"),
-            ),
-            (
-                "unassessed_axes_recorded",
-                ObservationBasis::Complete,
-                DeclaredToolAnnotations {
-                    read_only: None,
-                    destructive: None,
-                    idempotent: Some(true),
-                    open_world: Some(false),
-                },
-                "github.add_deploy_key",
-                json!({"owner": "acme", "repo": "prod-app"}),
-                Some("sha256:tooldigest-unassessed"),
-            ),
-            (
-                "incomplete_manifest_inconclusive",
-                ObservationBasis::Incomplete,
-                DeclaredToolAnnotations::default(),
-                "github.add_deploy_key",
-                json!({"owner": "acme", "repo": "prod-app"}),
-                None,
-            ),
-        ];
-
-        cases
-            .iter()
-            .map(|(case, basis, declared, tool, args, tool_digest)| {
-                json!({
-                    "case": case,
-                    "record": build_tool_annotation_conformance_record(
-                        *basis,
-                        declared,
-                        tool,
-                        *tool_digest,
-                        args
-                    ),
-                })
-            })
-            .collect()
-    }
-
-    fn contract_document() -> Value {
-        json!({
-            "schema_contract": TOOL_ANNOTATION_CONFORMANCE_SCHEMA,
-            "generated_by": "assay crates/assay-mcp-server proxy::annotation_conformance::build_tool_annotation_conformance_record (tool_annotation_conformance_contract_fixture)",
-            "note": "Canonical producer output, regenerated from build_tool_annotation_conformance_record. Consumers vendor this file verbatim. Regenerate with ASSAY_UPDATE_GOLDEN=1.",
-            "records": contract_records(),
-        })
-    }
-
-    #[test]
-    fn tool_annotation_conformance_contract_fixture() {
-        let generated = contract_document();
-        let path = contract_fixture_path();
-
-        if std::env::var("ASSAY_UPDATE_GOLDEN").is_ok() {
-            std::fs::create_dir_all(path.parent().unwrap()).unwrap();
-            let pretty = serde_json::to_string_pretty(&generated).unwrap();
-            std::fs::write(&path, format!("{pretty}\n")).unwrap();
-        }
-
-        let committed_text = std::fs::read_to_string(&path).unwrap_or_else(|_| {
-            panic!(
-                "missing {}; regenerate with ASSAY_UPDATE_GOLDEN=1",
-                path.display()
-            )
-        });
-        let committed: Value = serde_json::from_str(&committed_text).unwrap();
-        assert_eq!(
-            committed, generated,
-            "the committed tool-annotation conformance contract fixture is stale; regenerate with ASSAY_UPDATE_GOLDEN=1"
-        );
-
-        let records = generated["records"].as_array().unwrap();
-        assert_eq!(records.len(), 8);
-        for entry in records {
-            let rec = &entry["record"];
-            assert_eq!(rec["schema"], json!(TOOL_ANNOTATION_CONFORMANCE_SCHEMA));
-            let obj = rec.as_object().unwrap();
-            let mut keys: Vec<&str> = obj.keys().map(|k| k.as_str()).collect();
-            keys.sort_unstable();
-            assert_eq!(
-                keys,
-                [
-                    "assessed_axes",
-                    "conformance",
-                    "declared",
-                    "mismatch_kind",
-                    "non_claims",
-                    "observation_basis",
-                    "observed",
-                    "schema",
-                    "tool",
-                    "unassessed_axes"
-                ]
-            );
-            for forbidden in [
-                "decision",
-                "reason",
-                "forwarded",
-                "delivered",
-                "credential_alias",
-                "scopes",
-                "target_digest",
-                "caller_id",
-            ] {
-                assert!(
-                    rec.get(forbidden).is_none(),
-                    "carrier must not carry `{forbidden}`"
-                );
-            }
-            let text = serde_json::to_string(rec).unwrap();
-            for raw in ["ci-key", "alice@example.com"] {
-                assert!(
-                    !text.contains(raw),
-                    "raw argument value {raw} must not appear in the contract fixture"
-                );
-            }
-        }
-    }
-}
+mod tests;
